fix(palindrome/segment): overflow-safe count and midpoint
- OccurrenceCount.Count: widen accumulator + return to long (was: int,
  overflowed at n~65536 since total palindromes ~n^2/2). No callers.
- LazySegmentTree: all 8 midpoint sites use l + ((r - l) >> 1) instead
  of (l + r) >> 1 (overflow-safe for large l,r; consistent with
  ChairmanTree which already used the safe form).

+++ src/IAFahim.String.Palindrome/OccurrenceCount.cs
@@ -6,9 +6,9 @@ namespace IAFahim.String.Palindrome
     public static unsafe class OccurrenceCount
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int Count(byte* s, int n)
+        public static long Count(byte* s, int n)
         {
-            int count = 0;
+            long count = 0;
             int* odd = stackalloc int[n];
             Manacher.Odd(s, n, odd);
             for (int i = 0; i < n; i++)
